backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
python-telegram-bot
reportlab
python-dotenv
google-generativeai


